Laboratório VI/requirements.txt
paho-mqtt
Events